backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
pinecone-client==2.2.4
openai==1.3.0
sentence-transformers==2.2.2
numpy==1.24.3
python-multipart==0.0.6
websockets==12.0
aiofiles==23.2.1
